dev-lang/rust: reimport rust-1.41.1 from smauel overlay
new firefox-78esr demands at least rust-1.41, so providing it

upstreams binaries are built with llvm-9, which will most likely be on life support before the new esr cycle has ended. smauel has patches provided to build with llvm-10 and gcc-10 as well, so that should be good enough for a longer time.

this release has support for neon and thumb on armv7 at compile time, firefox demands these extra crates for it's own useflag. Todo: try to convince rust.configure of the new musl+neon nickname

todo: test with python-3.8 and enable if possible, maybe test bootstrapping with clang

Signed-off-by: Steffen Kuhn <nielson2@yandex.com>